+++ out/extension.js
@@ -20573,6 +20573,9 @@ var FeedbackViewProvider = class {
         this._connectBridge(webviewView);
       }
     });
+    if (webviewView.visible) {
+      this._connectBridge(webviewView);
+    }
     webviewView.onDidDispose(() => {
       this._view = null;
       this._bridge?.dispose();